changed parameter type

+++ struct/MultipleChoice.java
@@ -24,7 +24,7 @@ public class MultipleChoice extends QuizItem {
      * @param answerFour   the fourth answer option
      * @param correctIndex the correct index of the answer in the array
      */
-    public MultipleChoice(int id, int frequency, String question, String answerOne, String answerTwo,
+    public MultipleChoice(int id, double frequency, String question, String answerOne, String answerTwo,
             String answerThree,
             String answerFour, int correctIndex) {
         super(id, frequency);
